requirements.txt
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
colorzero==1.1
distro==1.5.0
gpiozero==1.6.2
gyp==0.1
idna==2.10
Mopidy==3.2.0
Mopidy-MPD==3.2.0
Mopidy-Spotify==4.1.1
playsound==1.3.0
pycparser==2.21
PyGObject==3.38.0
Pykka==2.0.3
pyspotify==2.1.3
python-apt==2.2.1
requests==2.25.1
RPi.GPIO==0.7.0
six==1.16.0
spidev==3.5
ssh-import-id==5.10
tornado==6.1
urllib3==1.26.5